🐛 Fix: 페이지에서 import 해 사용할 때 오류나는 것 해결
prop을 주고 올바른 타입 지정
