+++ src/types/card.ts
@@ -1,0 +1,23 @@
+export interface ResponseData {
+  likeCount: number;
+  tags: [
+    {
+      name: string;
+      id: number;
+    },
+  ];
+  writerId: number;
+  referenceUrl: string;
+  referenceTitle: string;
+  author: string;
+  content: string;
+  id: number;
+  isLiked: boolean;
+}
+
+export interface TagResponse {
+  tags: {
+    name: string;
+    id: number;
+  }[];
+}
