+++ src/types/sample.ts
@@ -35,20 +35,3 @@ export interface ButtonProps {
   onClick: () => void;
   variant?: 'primary' | 'secondary';
 }
-
-export interface ResponseData {
-  likeCount: number;
-  tags: [
-    {
-      name: string;
-      id: number;
-    },
-  ];
-  writerId: number;
-  referenceUrl: string;
-  referenceTitle: string;
-  author: string;
-  content: string;
-  id: number;
-  isLiked: boolean;
-}
